[TASK] Add RenderOfficialDocsFirsttime. Cool stuff!
This is "cool stuff" that creates a whole TYPO3 Sphinx Documentation
project from one 'manual.sxw' in just one step! I've just used
it the way it is now on the server to create the ReST files and
documentation for the TyposcriptReference. It works!
See RenderOfficialDocsFirsttime/README.txt for more information.

Wow! :-) :-) :-)

Change-Id: Ie8a146885e930e2c9a461ab8ef6894d1a729102f