[TASK] Add Yaml configuration features to t3sphinx
Our Python package 't3sphinx' in 'ExtendingSphinxForTYPO3' is extended.
We add the ability to Sphinx to read a GlobalSettings.yml and a
local Settings.yml file on top of the normal settings given in conf.py.
Both are in Yaml format.

Please see issue http://forge.typo3.org/issues/39794 for details.

Change-Id: I5e8dd59e82a7057f304cbf22f33fd9df57eb4491
Resolves: #39794