[!!!][TASK] New: T3PythonDocBuilderPackage
The "T3PythonDocBuilder" is an application that can be run on the
commandline on Linux, Mac and Windows like this:

  $ cd T3PythonDocBuilder
  $ python t3pdb_sxw2html.py  manual.sxw  tempdir

The infile "manual.sxw" is read. "tempdir" must specify a writable
folder. The application will parse the OpenOffice document and create
ReST files which are then rendered as HTML again.

Additionally a ready-to-run Sphinx documentation project ist created.

For details see README.rst.

Change-Id: I8f86beaa649305e4cbe348a26123def5401dc071
Resolves: #39828