[TASK] Add format string today_fmt to GlobalSettings.yml
Sphinx knows a placeholder "|today|" that can be part of the ReST source
of a manual. It's replaced by the current date at the time of rendering.
It seems to be useful to add::

  :Rendered: |today|

to the meta data. By default there is no time given so. Here we are
adding as configuration variable::

  today_fmt: '%Y-%m-%d %H:%M'

to GlobalSettings.py. This will display a "|today|" in the form
"2012-09-26 22:19".

You have to reinstall the "t3sphinx" Python package if you want to have
that in effect locally at home::

  $ cd .../RestTools.git/ExtendingSphinxForTYPO3
  $ python setup.py install

Change-Id: I1b66c9c1d77ea27641495859b6812bdb505debec
Resolves: #41424