[TASK] t3sphinx.ext.targets extension to list :ref:`TARGETS`
In our documents we define so called "labels" that afterwards serve as
symbolic name for link targets in intersphinx linking. It would be very
helpful for editors to see an overview of the labels that are defined in
a documentation project.

Here we introduce a new directive which does exactly that. It inserts a
list of ref targets into the document. The directive doesn't know any
options or parameters:

  .. ref-targets-list::

See http://forge.typo3.org/issues/42668

Change-Id: I983bd635b2d668e4a8c8b674ff992b735b87e5d8
Resolves: #42668