[BUGFIX] Update Intersphinx mapping
This is a follow-up to a previous patch changing the reference to
the Admin Guide to a reference to the Maintenance Guide. However
although the URL was change, the Intersphinx key was forgotten.
It must be changed too, especially since we have no links to that
manual yet.

Resolves: #48508
Releases: 1.0
Related: I903425309cbc448806aaeb2d3ece938b1c26716b

Change-Id: I7242b290fafc621fea494e8ce3c6940ffb2b884b
Reviewed-on: https://review.typo3.org/21062
Reviewed-by: Francois Suter
Tested-by: Francois Suter